Refactor LLM providers into package
